% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_function.R
\name{download_function}
\alias{download_function}
\title{A download Function}
\usage{
download_function(download = TRUE)
}
\arguments{
\item{download}{Do you want to download? Defaults to TRUE.}
}
\description{
This function allows you to download data from GEO
}
\examples{
download_function()
}
\keyword{download}

